requirements.txt
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
cython-bbox @ git+https://github.com/yanfengliu/cython_bbox.git@e7b95b6dc0933b10bd0ec02b90755a12ca996c68
distro==1.5.0
easydict==1.9
grpcio==1.33.2
grpcio-tools==1.33.2
idna==2.10
joblib==0.17.0
jsonpatch==1.26
jsonpointer==2.0
kiwisolver==1.3.1
matplotlib==3.3.3
munkres==1.1.4
natsort==7.0.1
numpy==1.19.4
opencv-python==4.4.0.46
packaging==20.4
Pillow==8.0.1
protobuf==3.14.0
pycocotools==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.0
scikit-build==0.11.1
scikit-learn==0.23.2
scipy==1.1.0
six==1.15.0
sklearn==0.0
tensorboardX==2.1
terminaltables==3.1.0
threadpoolctl==2.1.0
timm==0.1.20
torch==1.1.0
torchfile==0.1.0
torchvision==0.3.0
tornado==6.1
tqdm==4.52.0
urllib3==1.26.2
visdom==0.1.8.9
websocket-client==0.57.0
